infra: address PR review feedback
- Rename constants.py to gate_mappings.py
- Move single-consumer constants to their consumer modules
- Extract compilation pipeline into compilation.py
- Remove outdated TODO comment
- Fix test method name
- Remove unused Interpreter import from qasm_context.py
- Restore full docstrings
- Make QiskitProgramContext, SubstitutedTarget, compile_circuits public
